requirements.txt
easydict
joblib
numpy
opencv-python
Pillow
tensorflow
PyYAML
